services/chat_api/requirements.txt
# Core application
fastapi~=0.111.0
uvicorn[standard]~=0.29.0

# Database
sqlalchemy[asyncio]~=2.0.29
asyncpg~=0.29.0

# GCP & OpenAI
google-cloud-secret-manager~=2.19.0
google-cloud-aiplatform>=1.67.0

# Configuration and Utilities
pydantic-settings~=2.2.0
python-dotenv>=1.0.0

# Authentication (pin to compatible versions)
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0

# Explicitly set compatible version for FastAPI dependency
python-multipart>=0.0.7

requests==2.31.0

